Split agent subcommand metadata from heavy command handlers so search-code --help and agent command dispatch avoid loading the full commands chunk
Result: {"status":"keep","total_ms":146.091}
