+++ autoresearch.jsonl
@@ -1,3 +1,4 @@
 {"run":1,"commit":"485db5a","metric":0,"metrics":{},"status":"crash","description":"Tried to benchmark current working tree in ../coding-agents-plugins, but ./autoresearch.sh was not executable (permission denied).","timestamp":1776386354540,"segment":0,"confidence":null,"iterationTokens":null,"asi":{"hypothesis":"Use the existing harness to benchmark the current working tree before further optimization.","rollback_reason":"Benchmark command failed immediately because the harness lacks execute permission in this repo clone.","next_action_hint":"Invoke the harness via bash autoresearch.sh so experiments can proceed without changing file permissions.","error":"bash: ./autoresearch.sh: Permission denied"}}
 {"type":"config","name":"Speed up octocode-cli help startup","metricName":"total_ms","metricUnit":"ms","bestDirection":"lower"}
 {"run":2,"commit":"4f1c7f8","metric":613.193,"metrics":{},"status":"keep","description":"Baseline for built octocode-cli lightweight startup flows (--help, search-code --help, --version)","timestamp":1776386509616,"segment":1,"confidence":null,"iterationTokens":119,"asi":{"hypothesis":"Establish a baseline for representative built CLI help/version invocations before optimizing startup imports.","secondary_metrics":{"help_ms":203.97,"command_help_ms":181.953,"version_ms":227.269,"stdout_bytes":41535,"passes":1},"learned":"Even no-network help/version paths cost ~180-227ms each, suggesting startup/import overhead dominates more than command logic."}}
+{"run":3,"commit":"b6ce7a1","metric":266.728,"metrics":{},"status":"keep","description":"Code-split the built CLI and lazy-load commands/tool/help modules so lightweight help/version flows avoid eager startup of the full command graph","timestamp":1776386605093,"segment":1,"confidence":null,"iterationTokens":5504,"asi":{"hypothesis":"The main entrypoint eagerly imports heavy command and tool modules; code splitting plus lazy imports should cut no-network startup time without changing output.","secondary_metrics":{"help_ms":39.769,"command_help_ms":192.438,"version_ms":34.521,"stdout_bytes":41535,"passes":1},"learned":"Single-file bundling was the dominant cost. Splitting shrank the entry bundle from ~2.8MB to ~4KB and made general help/version dramatically faster, while command-specific help remained dominated by loading the commands chunk."}}
